Rename controller file to camelCase
